Update create-dmg to work in macOS Catalina 10.15
