+++ src/MacVim/dmg/README.md
@@ -1,0 +1,1 @@
+This folder contains resources used for creating the dmg installation file.
