+++ src/MacVim/dmg/background.png
