+++ src/MacVim/dmg/background.pxd
