+++ src/create-dmg/README
@@ -1,2 +1,1 @@
-Content of this directory (except background.png) is copied from
-https://github.com/andreyvit/create-dmg (5acf22f).
+Content of this directory is copied from https://github.com/create-dmg/create-dmg (v1.0.8 / 0985398).
