+++ src/create-dmg/tests/007-space-in-dir-name/my files/hello.txt
@@ -1,0 +1,1 @@
+Hello world
